requirements.txt
streamlit==1.36.0
Pillow==10.3.0
numpy>=1.26.0
opencv-python==4.9.0.80
ultralytics==8.1.14
tensorflow==2.18.0
